refactor: :recycle: refactor project folder structure
