--- components/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "conventionalCommits.scopes": [
-        "head",
-        "hero",
-        "about",
-        "work-experience"
-    ]
-}
