+++ src/components/About/About.tsx
@@ -1,6 +1,6 @@
 import { motion } from 'framer-motion';
 import React from 'react'
-import styles from '../styles/About.module.scss';
+import styles from './About.module.scss';
 
 type Props = {}
