+++ src/components/Hero/Hero.tsx
@@ -1,8 +1,8 @@
 import Link from "next/link";
 import React from "react";
 import { Cursor, useTypewriter } from "react-simple-typewriter";
-import BackgroundCircles from "./BackgroundCircles";
-import styles from "../styles/Hero.module.scss";
+import BackgroundCircles from "../BackgroundCircles/BackgroundCircles";
+import styles from "./Hero.module.scss";
 import Image from "next/image";
