+++ src/components/WorkExperience/WorkExperience.module.scss
@@ -1,4 +1,4 @@
-@import "./shared.module.scss";
+@import "../../styles/shared.module.scss";
 
 .container {
     height: 100vh;
